Smarter frontControllerDirectory method
The previous method simply returned `web` if there was a file named `craft` in the $sitePath (which means we’re running Craft 3).

The problem is that many sites, especially upgraded ones, will be serving out of `public`. This is also in line with the default Forge/ServerPilot default dirs.

So we test to see if either `/web` or `/public` exist, and return them if so. Otherwise default to the current method.

Signed-off-by: Andrew Welch <andrew@nystudio107.com>