Add link to upgrade instructions in outdated error
Previously when checking if the Valet version in use was the latest,
you'd have to go to Valet repo on GitHub, find out that all the docs
were on laravel.com, go there, find the upgrade instructions, and then
upgrade.

This can be simplified by giving the user the option to view the link.